Adds authorization check

+++ app/Http/Controllers/ListingController.php
@@ -53,6 +53,11 @@ class ListingController extends Controller
    }
 
    public function update(Request $request, Listing $listing) {
+
+    if($listing->user_id != auth()->id()) {
+        abort('403', 'Not authorized.');
+    }
+
     $formFields = $request->validate([
         'title' => 'required',
         'company' => ['required'],
@@ -74,7 +79,14 @@ class ListingController extends Controller
    }
 
    public function destroy(Listing $listing){
+    if($listing->user_id != auth()->id()) {
+        abort('403', 'Not authorized.');
+    }
     $listing->delete();
     return redirect('/')->with('message', 'Deleted listing successfully.');
+   }
+
+   public function manage(){
+    return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
    }
 }
